refactor: Move menu helper into menu module
