+++ src-tauri/src/main.rs
@@ -10,41 +10,13 @@ use mimalloc::MiMalloc;
 static GLOBAL: MiMalloc = MiMalloc;
 
 use std::sync::Arc;
-use tauri::{AppHandle, Emitter, Manager};
-use tauri_plugin_dialog::DialogExt;
+use tauri::Manager;
 
 use trendline_lib::pipeline::manager::ProcessingManager;
 use trendline_lib::state::capture::CaptureState;
 use trendline_lib::state::proxy_register::ProxyRegister;
 use trendline_lib::{commands, menu, proxy};
 
-fn handle_menu_event(app: &AppHandle, event: tauri::menu::MenuEvent) {
-    let window = app.get_webview_window("main").unwrap();
-    match event.id().as_ref() {
-        "open_recording" => {
-            app.dialog().file().pick_file(move |path_buf| {
-                if let Some(path) = path_buf {
-                    println!("File selected for opening: {}", path.to_string());
-                    window.emit("file-opened", path.to_string()).unwrap();
-                }
-            });
-        }
-        "save_recording" => {
-            // Using the new dialog plugin API
-            app.dialog()
-                .file()
-                .add_filter("Trendline Recording", &["json"])
-                .set_file_name("recording.json")
-                .save_file(move |path_buf| {
-                    if let Some(path) = path_buf {
-                        println!("File selected for saving: {}", path.to_string());
-                    }
-                });
-        }
-        // ... other handlers remain the same
-        _ => {}
-    }
-}
 
 fn main() {
     tauri::Builder::default()
@@ -53,12 +25,12 @@ fn main() {
         .plugin(tauri_plugin_clipboard_manager::init())
         .plugin(tauri_plugin_dialog::init())
         .setup(|app| {
+            let app_handle = app.handle().clone();
+
             let menu = menu::create_app_menu(app.handle())?;
             app.set_menu(menu)?;
-
-            let app_handle = app.handle().clone();
             app.on_menu_event(move |_app, event| {
-                handle_menu_event(&app_handle, event);
+                menu::handle_menu_event(&app_handle, event);
             });
 
             let capture = CaptureState::new();
